requirements.txt
requests
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
